agent_bot/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
mysql-connector-python
fastapi 
uvicorn
dotenv